pyproject.toml
[tool.poetry]
name = "airbnb-clone-django"
version = "0.1.0"
description = "airbnb clone with django and react"
authors = ["toweringcloud <toweringcloud@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.12"
django = "^5.1.1"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
pyjwt = "^2.9.0"
django-cors-headers = "^4.5.0"
black = "^24.8.0"
django-extensions = "^3.2.3"
ipython = "^8.29.0"
Werkzeug = "^3.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
